First commit of PeakFinder. This tool finds mountains with high prominence, nearby roads, and paths between them of a specified length.
First, this tool uses the Summits On The Air (SOTA) ham radio database to find peaks with high prominence.
Second, this tool downloads Open Street Map (OSM) data for the immediate vicinity of the peak and looks for roads.
Third, this tool looks for permutations and combinations of peaks with roads that are the specified distance range apart.
Fourth [UNIMPLEMENTED] this tool will compute the terrain profile along possible paths and check for line of sight.

Downloading OSM data is slow, so all of the needed files are part of this commit.